Fix regeneration
The RISC bootstrap regen script now only copies a new copy of the RISC
code to the bootstrap directory if it is changed. This means that the
timestamp of the risc_bootstrap.txt isn't always changing, speeding up
the compilaiton of the compiler (which doesn't think the compiler always
needs to be regenerated).

Thus the Makefile now auto-regenerates, which is a little slower than
necessary on the first compilation, but otherwise works correctly

+++ risc_bootstrap/regen.sh
@@ -4,4 +4,6 @@ set -e
 
 cd ../compiler
 ../build/rcompile Compiler.ob
-cp ../build/risc_code.txt ../risc_bootstrap
+if ! cmp -s ../build/risc_code.txt ../risc_bootstrap/risc_code.txt; then
+  cp ../build/risc_code.txt ../risc_bootstrap
+fi
